♻️ 계정 설정 페이지에서 회원가입 페이지로 변경
initialize-account -> signup
